Try to switch to zsh if available
It's a bit meh really but whatcha gonna do. We don't want to prompt for
passwords or anything, so we don't use chsh but just exec one shell from
another. Also we don't want to have to run package managers because I
guess we might already be non-root at this point?